requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
django-environ
psycopg2-binary
redis
celery[redis]
httpx
anthropic>=0.40.0
groq>=1.0.0
chromadb
sentence-transformers
numpy
pandas
gunicorn
